p2m_kernels*: atomicAdd fixed for all CUDA versions
Somehow the previous fix with commit cc62e0e3 did not work
any more on the P100 for CUDA v8 although it had been tested.
This patch now works for both CUDA v<8 and v==8 and any architecture.

FYI:
With this patch, we use double precision atomicAdd either with
CUDA v8 as defined in the library or with CUDA v<8 we define it
via software emulation (slower than hardware acceleration).
From Pascal architecture on, the GPUs support
hardware accelerated double precision atomicAdd with CUDA v8.
PyPIC on the GPU makes use of this now.