GPU Meshing: rename RectMesh25D to SlicingMesh25D
--> reason: in the PyPIC_GPU implementation, the fast kernels
include the 2D and 3D meshing by checking whether the dimension is 2 or 3
and whether the mesh class contains "RectMesh". 2.5D is not implemented
in the fast kernels (it's not usually used anyway).
By renaming the mesh to SlicingMesh25D, it will be properly called
via the (slower) python functions, which differ from the 3D version.
Previous to this bug fix, the 2.5D mesh would lead to the 3D mesh's
longitudinal linear interpolation along the slices because the
fast kernel was used (and it should not).